Blue added
I have added the other color, BUT there is a problem bc when the blue is picked it does not change into purple I have to change it to white (switch it off) and then I can change it to purple. BUT when I am in the purple mode I can change it to blue. I understand that the script  is loaded from top to bottom :).